move new tests to existing axes_tests module under builder dir
